python/requirements.txt
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
dataclasses-json==0.6.2
distro==1.8.0
exceptiongroup==1.1.3
faiss-cpu==1.7.4
filelock==3.13.1
frozenlist==1.4.0
fsspec==2023.10.0
h11==0.14.0
html5tagger==1.3.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.1
idna==3.4
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.335
langsmith==0.0.64
llvmlite==0.41.1
MarkupSafe==2.1.3
marshmallow==3.20.1
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.1
openai==1.2.3
openai-whisper @ git+https://github.com/openai/whisper.git@fcfeaf1b61994c071bba62da47d7846933576ac9
package-name==0.1
packaging==23.2
pandas==2.1.3
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
rank-bm25==0.2.2
regex==2023.10.3
requests==2.31.0
sanic==23.6.0
sanic-routing==23.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
sympy==1.12
tenacity==8.2.3
tiktoken==0.5.1
torch==2.0.0
tqdm==4.66.1
tracerite==1.1.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
urllib3==2.0.7
uvloop==0.19.0
websockets==12.0
yarl==1.9.2
